Implement url validity check UI on desktop
