+++ desktopApp/src/jvmMain/kotlin/com/prof18/feedflow/navigation/Screen.kt
@@ -3,6 +3,8 @@ package com.prof18.feedflow.navigation
 import com.arkivanov.essenty.parcelable.Parcelable
 
 sealed class Screen : Parcelable {
-    object Home : Screen()
-    object FeedList : Screen()
+    data object Home : Screen()
+    data object FeedList : Screen()
+
+    data object ImportExport : Screen()
 }
