+++ shared/src/commonMain/kotlin/com/prof18/feedflow/presentation/preview/PreviewItems.kt
@@ -2,11 +2,11 @@
 
 package com.prof18.feedflow.presentation.preview
 
+import com.prof18.feedflow.core.model.FeedImportExportState
 import com.prof18.feedflow.core.model.FeedItem
 import com.prof18.feedflow.core.model.FeedSource
 import com.prof18.feedflow.core.model.ParsedFeedSource
 import com.prof18.feedflow.domain.model.Browser
-import com.prof18.feedflow.presentation.model.FeedImportExportState
 
 val feedItemsForPreview = listOf(
     FeedItem(
